Added method to timeservice that counts minutes since certain time, created enum for resource type and implemented DB updates to resources

+++ DotNetTribes/DTOs/ResourceDto.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using DotNetTribes.Enums;
 
 namespace DotNetTribes.DTOs
 {
@@ -6,6 +7,6 @@ namespace DotNetTribes.DTOs
     {
         public string Type { get; set; }
         public int Amount { get; set; }
-        public long UpdatedAt { get; set; }
+        public int UpdatedAt { get; set; }
     }
 }
